shutdown-daemon/requirements.txt
bcrypt==4.0.1
cryptography==41.0.3
python-dotenv
requests
pystray>=0.19.4
pillow>=9.0.0